data program/data_analysis/requirements.txt
flask==3.0.1
cassandra-driver==3.28.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0